Add points to the edge of image
I've created a shortcut (Q) to add a polygon point pressing the Q on keyboard. This is useful when you want to add point polygons to the edge of the image. Using only the mouse is awful because the mouse point should be inside the image. This is hard to do when you are in the margins.